+++ plutarch-ledger-api/src/Plutarch/LedgerApi/V1/Time.hs
@@ -7,6 +7,7 @@ module Plutarch.LedgerApi.V1.Time (
 
 import Plutarch.Builtin (PDataNewtype (PDataNewtype))
 import Plutarch.DataRepr (DerivePConstantViaData (DerivePConstantViaData))
+import Plutarch.Enum (PCountable (psuccessor, psuccessorN))
 import Plutarch.LedgerApi.Utils (Mret)
 import Plutarch.Lift (
   PConstantDecl,
@@ -39,6 +40,15 @@ newtype PPosixTime (s :: S) = PPosixTime (Term s (PDataNewtype PInteger))
     , -- | @since 2.0.0
       PShow
     )
+
+-- | @since WIP
+instance PCountable PPosixTime where
+  {-# INLINEABLE psuccessor #-}
+  psuccessor = phoistAcyclic $ plam (+ 1)
+  {-# INLINEABLE psuccessorN #-}
+  psuccessorN = phoistAcyclic $ plam $ \p t ->
+    let p' = pcon . PPosixTime . pcon . PDataNewtype . pdata . pto $ p
+     in p' + t
 
 -- | @since 2.0.0
 instance PIntegral PPosixTime where